Revert "Fix the virtual media creation for core/gallery block. Resolve virtua…" (#1943)
This reverts commit c5f511a08508f54535499dc546877f7b942d987c.
